requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.0
requests==2.32.0
lxml==5.1.0
python-dateutil==2.9.0
python-dotenv==1.0.1
